backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
google-generativeai>=0.8.0
supabase>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
openai>=1.0.0
groq>=0.4.0
